+++ baklava/src/interfaces.ts
@@ -1,0 +1,3 @@
+export interface Env {
+  rooms: DurableObjectNamespace;
+}
