--- baklava/src/room/interfaces.ts
@@ -1,3 +1,0 @@
-export interface Env {
-  rooms: DurableObjectNamespace; // TODO types
-}
