+++ baklava/wrangler.toml
@@ -1,6 +1,6 @@
 #:schema node_modules/wrangler/config-schema.json
 name = "baklava"
-main = "src/room/index.ts"
+main = "src/index.ts"
 compatibility_date = "2024-08-06"
 
 # Automatically place your workloads in an optimal location to minimize latency.
